Added easier way of selecting application key
It is not possible to include all application key in the main configuration field like this:

"key,key,key or name:key,name:key,name:key"

The first shows the keys in the action settings, the second shows the names.

WARNING
Using this will not break earlier settings but you will have to add (all) keys or name:key pairs into the configuration menu before you start editing the actions.

Signed-off-by: Marcel Trapman <marceltrapman@it2be.com>